+++ redwood-lazylayout-compose/src/commonMain/kotlin/app/cash/redwood/lazylayout/compose/LazyList.kt
@@ -39,16 +39,15 @@ internal fun LazyList(
 ) {
   val itemProvider = rememberLazyListItemProvider(content)
   val itemCount = itemProvider.itemCount
-  val itemsBefore = (state.firstIndex - state.preloadBeforeItemCount).coerceAtLeast(0)
-  val itemsAfter = (itemCount - (state.lastIndex + state.preloadAfterItemCount).coerceAtMost(itemCount)).coerceAtLeast(0)
+  val loadRange = state.loadRange(itemCount)
   val placeholderPoolSize = 20
   LazyList(
     isVertical = isVertical,
     onViewportChanged = { localFirstVisibleItemIndex, localLastVisibleItemIndex ->
       state.onUserScroll(localFirstVisibleItemIndex, localLastVisibleItemIndex)
     },
-    itemsBefore = itemsBefore,
-    itemsAfter = itemsAfter,
+    itemsBefore = loadRange.first,
+    itemsAfter = itemCount - loadRange.count() - loadRange.first,
     width = width,
     height = height,
     margin = margin,
@@ -57,7 +56,7 @@ internal fun LazyList(
     scrollItemIndex = state.programmaticScrollIndex,
     placeholder = { repeat(placeholderPoolSize) { placeholder() } },
     items = {
-      for (index in itemsBefore until itemCount - itemsAfter) {
+      for (index in loadRange) {
         key(index) {
           itemProvider.Item(index)
         }
@@ -83,13 +82,12 @@ internal fun RefreshableLazyList(
 ) {
   val itemProvider = rememberLazyListItemProvider(content)
   val itemCount = itemProvider.itemCount
-  val itemsBefore = (state.firstIndex - state.preloadBeforeItemCount).coerceAtLeast(0)
-  val itemsAfter = (itemCount - (state.lastIndex + state.preloadAfterItemCount).coerceAtMost(itemCount)).coerceAtLeast(0)
+  val loadRange = state.loadRange(itemCount)
   val placeholderPoolSize = 20
   RefreshableLazyList(
     isVertical,
-    itemsBefore = itemsBefore,
-    itemsAfter = itemsAfter,
+    itemsBefore = loadRange.first,
+    itemsAfter = itemCount - loadRange.count() - loadRange.first,
     onViewportChanged = { localFirstVisibleItemIndex, localLastVisibleItemIndex ->
       state.onUserScroll(localFirstVisibleItemIndex, localLastVisibleItemIndex)
     },
@@ -104,7 +102,7 @@ internal fun RefreshableLazyList(
     placeholder = { repeat(placeholderPoolSize) { placeholder() } },
     pullRefreshContentColor = pullRefreshContentColor,
     items = {
-      for (index in itemsBefore until itemCount - itemsAfter) {
+      for (index in loadRange) {
         key(index) {
           itemProvider.Item(index)
         }