+++ redwood-lazylayout-compose/src/commonTest/kotlin/app/cash/redwood/lazylayout/compose/LazyListTest.kt
@@ -101,12 +101,12 @@ class LazyListTest {
 
       with(awaitSnapshot()) {
         val lazyList = single() as LazyListValue
-        assertThat(lazyList.itemsBefore).isEqualTo(35)
-        assertThat(lazyList.itemsAfter).isEqualTo(25)
+        assertThat(lazyList.itemsBefore).isEqualTo(50 - 5)
+        assertThat(lazyList.itemsAfter).isEqualTo(100 - (60 + 20))
         assertThat(lazyList.placeholder)
           .isEqualTo(List(20) { TextValue(Modifier, "Placeholder") })
         assertThat(lazyList.items)
-          .isEqualTo(List(40) { TextValue(Modifier, (it + 35).toString()) })
+          .isEqualTo(List(35) { TextValue(Modifier, (it + 45).toString()) })
       }
     }
   }
@@ -117,8 +117,7 @@ class LazyListTest {
       var index5ComposeCount = 0
       setContent {
         val lazyListState = rememberLazyListState().apply {
-          preloadBeforeItemCount = 0
-          preloadAfterItemCount = 0
+          preloadItems = false
         }
         LazyColumn(
           state = lazyListState,
@@ -155,7 +154,16 @@ class LazyListTest {
       awaitSnapshot()
       assertThat(index5ComposeCount).isEqualTo(1)
 
-      // But it's recomposed again when scrolled back on screen.
+      // The item at index 5 remains composed because (6, 10) is contiguous with (4, 8).
+      lazyList.onViewportChanged(4, 8)
+      awaitSnapshot()
+      assertThat(index5ComposeCount).isEqualTo(1)
+
+      // Scrolling to a non-contiguous range causes a recomposition.
+      lazyList.onViewportChanged(9, 13)
+      awaitSnapshot()
+      assertThat(index5ComposeCount).isEqualTo(1)
+
       lazyList.onViewportChanged(4, 8)
       awaitSnapshot()
       assertThat(index5ComposeCount).isEqualTo(2)